k8s host logic should prefer user-defined variable
`KUBERNETES_SERVICE_HOST` will always [1] be set in a typical
k8s environment, and so the initial `if` logic always takes
precedence, and doesn't allow the consideration of the `else if`,
in this case the variable `CLOUD_PROVIDER_K8S_KUBERNETES_HOST`.
We switch the logic around to make `KUBERNETES_SERVICE_HOST` be
the fallthrough choice, before erroring-out.

This supports environments where the `panosc-instances` namespace,
and thus the pods, are not running on the same k8s environment as
the portal microservices.

[1] https://github.com/hashicorp/vault/blob/master/serviceregistration/kubernetes/client/config.go#L13